pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clipthread-ui-tkinter"
version = "0.1.0"
description = "A basic tkinter UI for Clipthread"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["clipthread", "ui", "clipboard", "sync", "synchronization"]
authors = [
    { name = "Leo Cances", email = "leocances@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyperclip ~= 1.8.2",
    "aiohttp ~= 3.11.11",
    "pydantic ~= 2.10.4",
    "python-dotenv ~= 1.0.1",
    "cryptography ~= 44.0.0",
    "clipthread-server ~= 0.0.2",
    "sv_ttk ~= 2.6.0",
    "pynput ~= 1.7.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.981",
]

[project.urls]
Homepage = "https://github.com/clipthread/clipthread-ui-tkinter.git"
Repository = "https://github.com/clipthread/clipthread-ui-tkinter.git"
Documentation = "https://github.com/clipthread/clipthread-ui-tkinter#readme"
"Bug Tracker" = "https://github.com/clipthread/clipthread-ui-tkinter/issues"


[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["book.png", "clipthread-tkinter-ui.service"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[project.scripts]
clipthread-ui = "ui:start_ui"
clipthread-ui-install = "service_install:install_systemd_service"


